Rename auto.js to extend-expect.js
